[LibOS] Move Linux ABI headers to a separate directory (part 1)
