+++ libos/src/sys/libos_sched.c
@@ -267,8 +267,8 @@ long libos_syscall_sched_getaffinity(pid_t pid, unsigned int user_mask_size,
 
 /* Approx. implementation that returns a random bit that is set from the cpu affinity mask
  * associated with the current calling thread */
-long libos_syscall_getcpu(unsigned* cpu, unsigned* node, struct getcpu_cache* unused) {
-    __UNUSED(unused);
+long libos_syscall_getcpu(unsigned* cpu, unsigned* node, void* unused_cache) {
+    __UNUSED(unused_cache);
 
     int ret;
     if (cpu && !is_user_memory_writable(cpu, sizeof(*cpu)))
