+++ libos/src/vdso/arch/x86_64/vdso.h
@@ -7,4 +7,4 @@
 int __vdso_clock_gettime(clockid_t clock, struct timespec* t);
 int __vdso_gettimeofday(struct timeval* tv, struct timezone* tz);
 time_t __vdso_time(time_t* t);
-long __vdso_getcpu(unsigned* cpu, struct getcpu_cache* unused);
+long __vdso_getcpu(unsigned* cpu, void* unused_cache);
